requirements.txt
flask
easyocr
pillow
gunicorn
